Restore support for Debian 10 (buster)

+++ build-helper.sh
@@ -84,13 +84,21 @@ mk-build-deps -ir -t "apt-get -o Debug::pkgProblemResolver=yes -y --no-install-r
 
 # Build packages
 log "Building package with DEB_BUILD_OPTIONS set to '${DEB_BUILD_OPTIONS:-}'"
+debuild_args=
+# supported since Debian 11 (bullseye)
+if dpkg-buildpackage --sanitize-env --help &> /dev/null; then
+    debuild_args+=' --sanitize-env'
+fi
+
 BUILD_START_TIME="$EPOCHSECONDS"
 # supported since Debian 12 (bookworm)
 if unshare --map-users 1,1,100 --help &> /dev/null; then
-    unshare --user --map-root-user --net --map-users 1,1,100 --map-users 65534,65534,1 --map-groups 1,1,100 --map-groups 65534,65534,1 --setuid "$(id -u build-runner)" --setgid "$(id -g build-runner)" -- debuild --prepend-path /usr/lib/ccache --preserve-envvar CCACHE_DIR --sanitize-env -rfakeroot -b --no-sign -sa | tee "${CDEBB_BUILD_DIR}/build.log"
+    # shellcheck disable=SC2086
+    unshare --user --map-root-user --net --map-users 1,1,100 --map-users 65534,65534,1 --map-groups 1,1,100 --map-groups 65534,65534,1 --setuid "$(id -u build-runner)" --setgid "$(id -g build-runner)" -- debuild --prepend-path /usr/lib/ccache --preserve-envvar CCACHE_DIR -rfakeroot -b --no-sign -sa ${debuild_args} | tee "${CDEBB_BUILD_DIR}/build.log"
 else
     log "unshare(1) does not support --map-users, falling back to runuser(1); build has network access"
-    runuser -u build-runner -- debuild --prepend-path /usr/lib/ccache --preserve-envvar CCACHE_DIR --sanitize-env -rfakeroot -b --no-sign -sa | tee "${CDEBB_BUILD_DIR}/build.log"
+    # shellcheck disable=SC2086
+    runuser -u build-runner -- debuild --prepend-path /usr/lib/ccache --preserve-envvar CCACHE_DIR -rfakeroot -b --no-sign -sa ${debuild_args} | tee "${CDEBB_BUILD_DIR}/build.log"
 fi
 log "Build completed in $((EPOCHSECONDS - BUILD_START_TIME)) seconds"
 
@@ -112,7 +120,12 @@ if [ -n "${RUN_LINTIAN+x}" ]; then
     apt-get install -y --no-install-recommends lintian
     adduser --system --no-create-home lintian-runner
     log "+++ Lintian Report Start +++"
-    runuser -u lintian-runner -- lintian --display-experimental --info --display-info --pedantic --tag-display-limit 0 --color always --verbose --fail-on none "${CDEBB_BUILD_DIR}"/*.changes | tee "${CDEBB_BUILD_DIR}/lintian.log"
+    # supported since Debian 11 (bullseye)
+    if lintian --help | grep -F -w -- '\b--fail-on\b' &> /dev/null; then
+        runuser -u lintian-runner -- lintian --display-experimental --info --display-info --pedantic --tag-display-limit 0 --color always --verbose --fail-on none "${CDEBB_BUILD_DIR}"/*.changes | tee "${CDEBB_BUILD_DIR}/lintian.log"
+    else
+        runuser -u lintian-runner -- lintian --display-experimental --info --display-info --pedantic --tag-display-limit 0 --color always --verbose "${CDEBB_BUILD_DIR}"/*.changes | tee "${CDEBB_BUILD_DIR}/lintian.log"
+    fi
     log "+++ Lintian Report End +++"
 fi
